% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageGeneBodyScaled.R
\name{axisGeneBodyScaled}
\alias{axisGeneBodyScaled}
\title{axisGeneBodyScaled}
\usage{
axisGeneBodyScaled(
  margin_outer = 500,
  margin_inner = 500,
  genebody_scaler = 1000,
  add_line = TRUE
)
}
\description{
ass x-axis to composite plot of gene body scaled matrix
}
