% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binMeansMatrix.R
\name{binMeansMatrix}
\alias{binMeansMatrix}
\title{binMeansMatrix}
\usage{
binMeansMatrix(my_mat, my_binning = 10)
}
\arguments{
\item{my_mat}{input matrix.}

\item{my_binning}{integer of bin size.}
}
\value{
output matrix with original number of rows and \code{my_binning} times less number of columns.
}
\description{
calculates the mean of non-overlapping bins for each row in a matrix.
}
\examples{
m <- matrix(1:100, ncol=10)
binMeansMatrix(m, 2)
}
\seealso{
\code{\link[matrixStats]{binMeans}}
}
