% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReadsOverGrangesInBAM.R
\name{countReadsOverGrangesInBAM}
\alias{countReadsOverGrangesInBAM}
\title{countReadsOverGrangesInBAM}
\usage{
countReadsOverGrangesInBAM(
  my_ranges,
  bam_file,
  bam_type = c("SINGLE", "PAIRED"),
  subsampling = NULL,
  frag.min = 0,
  frag.max = 500,
  frag.length = 150
)
}
\description{
count the number of reads over genomic ranges in a BAM file
}
