% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepBSgenomeSequences.R
\name{keepBSgenomeSequences}
\alias{keepBSgenomeSequences}
\title{keepBSgenomeSequences}
\usage{
keepBSgenomeSequences(genome, seqnames)
}
\description{
subset BS genome by chromosome(s)
}
\details{
Source: http://supportupgrade.bioconductor.org/p/83588/#83594
}
