% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderMats.R
\name{orderMats}
\alias{orderMats}
\title{orderMats}
\usage{
orderMats(my_mats, my_order = 1:nrow(get(my_mats[1])))
}
\arguments{
\item{my_mats}{Character vector of matrix names.}

\item{my_order}{Integer vector of the matrix row order.}
}
\description{
order or filter matrices by a vector
}
