% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComposite.R
\name{plotComposite}
\alias{plotComposite}
\title{plotComposite}
\usage{
plotComposite(
  my_sample_mats,
  my_sub_range = 1:400,
  ylims = c(1, 5),
  my_binning = 10,
  my_colors_composite,
  my_title = "",
  yaxt = "s",
  site_label = 0,
  add_line = FALSE,
  line_lwd = 2,
  smoother = 11,
  log_scale = FALSE,
  add_range = c(NULL, "SD", "SEM", "CI", "IQR"),
  add_axis = TRUE
)
}
\description{
plot composite signal
}
