% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap}
\usage{
plotHeatmap(
  my_sample_mats,
  my_sample_names = "",
  my_site_name = "0",
  font_size = 0.75,
  my_colors = colorRampPalette(brewer.pal(9, "Blues"))(100),
  min_value = 1,
  max_value = 5,
  my_binning = 10,
  smoother = 1
)
}
\description{
plot heatmap signal
}
