% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLegendtoPosition.R
\name{plotLegendtoPosition}
\alias{plotLegendtoPosition}
\title{plotLegendtoPosition}
\usage{
plotLegendtoPosition(
  conditions,
  legend_colors,
  legend_size = 0.8,
  horizontal = T,
  position = c(0, 0.25)
)
}
\description{
plot legend to positon of x and y coordinates on figure
}
