% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRanges.R
\name{plotRanges}
\alias{plotRanges}
\title{plotRanges}
\usage{
plotRanges(
  my_ranges,
  my_region,
  my_color,
  my_border = my_color,
  my_ylims = c(0, 1),
  my_title = "Track"
)
}
\description{
plots rectangles representing genomic ranges (e.g. peaks)
}
