% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVennOverlaps.R
\name{plotVennOverlaps}
\alias{plotVennOverlaps}
\title{plotVennOverlaps}
\usage{
plotVennOverlaps(
  my_granges_names,
  my_labels = NULL,
  my_venn_colors = NULL,
  line_width = 5
)
}
\arguments{
\item{my_granges_names}{character vector of granges object names.}

\item{my_labels}{character vector of label names.}

\item{my_venn_colors}{character vector of colors.}

\item{line_width}{integer of line thickness.}
}
\description{
plots Venn diagram of overlapping genomic ranges (e.g. peaks)
}
