% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageGeneBodyScaled.R
\name{resizeVector}
\alias{resizeVector}
\title{resizeVector}
\usage{
resizeVector(x, out_length)
}
\description{
resizes a vector \code{x} to a length of \code{out_length} using spline interpolation.
}
