% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRangesToFasta.R
\name{writeRangesToFasta}
\alias{writeRangesToFasta}
\title{writeRangesToFasta}
\usage{
writeRangesToFasta(my_genome, my_ranges, my_file_path)
}
\arguments{
\item{my_genome}{BSgenome object.}

\item{my_ranges}{genomic ranges object.}

\item{my_file_path}{output directory.}
}
\description{
writes sequences of genomic ranges into fasta file and ranges into bed file.
}
