% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReadsOverGrangesInBAM.R
\name{countReadsOverGrangesInBAM}
\alias{countReadsOverGrangesInBAM}
\title{countReadsOverGrangesInBAM}
\usage{
countReadsOverGrangesInBAM(
  my_ranges,
  bam_file,
  bam_type = c("SINGLE", "PAIRED"),
  single_read.unique = FALSE,
  frag.min = 0,
  frag.max = 500,
  frag.length = 150,
  frag.unique = c(NULL, "oneEND", "bothEND"),
  subsampling = c(NULL, 1e+06)
)
}
\description{
count the number of reads over genomic ranges in a BAM file
}
